refactor(transcript): move split_system from context.render to transcript
split_system is a structural transcript primitive (like safe_window /
entries_to_messages), not a render-specific helper — and the previous commit
gave it an out-of-context consumer (runtime/loop.py reaching into
context.render). Move it to transcript.py, its natural home, so loop.py,
compaction.py, and render.py all import it from the foundational module they
already depend on. Pure relocation; no behavior change.
